added server side base player and game start code

+++ server/box_sv.cpp
@@ -1,7 +1,7 @@
 #include "box_sv.h"
 #include "net_sv.h"
 
-box_sv::box_sv(level *lvl, char abox_type, vec *pos, uint ahealth)
+box_sv::box_sv(level *lvl, char abox_type, vec *pos, float ahealth)
 : box(lvl, abox_type, pos, ahealth)
 {
 	//lvl_sv = lvl;
